修复提成订单

+++ bonus/src/main/java/com/jfeat/am/module/bonus/services/domain/dao/QueryBonusDao.java
@@ -80,7 +80,7 @@ public interface QueryBonusDao {
 
     Float queryProportion(@Param("userId") Long userId);
 
-    Float queryOrderAmount(@Param("userId") Long userId);
+    BigDecimal queryOrderAmount(@Param("userId") Long userId);
 }
